refactor: improve data structure for allWords in data layer
Commit description:
I made improvements to the data structure of the allWords variable in the data layer. Previously, allWords was a setOf words, but I have now refactored it into a Map. This change allows for dynamic retrieval of random words and their respective keys, enabling the application to display hints dynamically on the screen.

Changes made:

Refactored the allWords variable from a setOf words to a Map data structure.
Implemented the necessary logic to retrieve random words and their respective keys from the Map.
Improved the functionality of the application to dynamically display hints on the screen.