chore: update value of PLAY_AGAIN_BUTTON_DIALOG string in brazilian Portuguese strings file
Commit description:
I made a minor adjustment to the value of the PLAY_AGAIN_BUTTON_DIALOG string in the Brazilian Portuguese strings file for layout purposes. The previous value was "Jogar novamente," but I have updated it to "Reiniciar."

Changes made:

Modified the value of the PLAY_AGAIN_BUTTON_DIALOG string in the Brazilian Portuguese strings file.
Updated the string value to "Reiniciar" for improved layout and alignment.